Move validation hooks to common/validation/
