+++ packages/core/src/common/hooks/index.ts
@@ -1,7 +1,2 @@
 export { Log, LogOptions } from './log.hook';
 export { UserRequired } from './user-required.hook';
-export { ValidateBody } from './validate-body.hook';
-export { ValidateCookie } from './validate-cookie.hook';
-export { ValidateHeader } from './validate-header.hook';
-export { ValidatePathParam } from './validate-path-param.hook';
-export { ValidateQueryParam } from './validate-query-param.hook';
