+++ packages/core/src/common/index.ts
@@ -2,3 +2,4 @@ export * from './encoding';
 export * from './hooks';
 export * from './tokens';
 export * from './utils';
+export * from './validation';
