+++ packages/core/src/common/validation/helpers/index.ts
@@ -1,0 +1,1 @@
+export { isFunction } from './is-function.util';
