+++ packages/core/src/common/validation/helpers/is-function.util.ts
