+++ packages/core/src/common/validation/validate-cookie.hook.spec.ts
