+++ packages/core/src/common/validation/validate-header.hook.spec.ts
