+++ packages/core/src/common/validation/validate-header.hook.ts
@@ -1,7 +1,5 @@
-// std
-import { ValidateFunction } from 'ajv';
-
 // FoalTS
+import { ValidateFunction } from 'ajv';
 import {
   ApiParameter,
   ApiResponse,
@@ -9,32 +7,33 @@ import {
   Hook,
   HookDecorator,
   HttpResponseBadRequest,
-  IApiQueryParameter,
+  IApiHeaderParameter,
   OpenApi,
   ServiceManager
 } from '../../core';
 import { getAjvInstance } from '../utils';
-import { isFunction } from './is-function.util';
+import { isFunction } from './helpers';
 
 /**
- * Hook - Validate a specific query parameter against an AJV schema.
+ * Hook - Validate a specific header against an AJV schema.
  *
  * @export
- * @param {string} name - Query parameter name.
+ * @param {string} name - Header name.
  * @param {(object | ((controller: any) => object))} [schema={ type: 'string' }] - Schema used to
- * validate the query parameter.
+ * validate the header.
  * @param {{ openapi?: boolean, required?: boolean }} [options={}] - Options.
  * @param {boolean} [options.openapi] - Add OpenApi metadata.
- * @param {boolean} [options.required] - Specify is the query is optional.
+ * @param {boolean} [options.required] - Specify is the header is optional.
  * @returns {HookDecorator} The hook.
  */
-export function ValidateQueryParam(
+export function ValidateHeader(
   name: string,
   schema: object | ((controller: any) => object) = { type: 'string' },
   options: { openapi?: boolean, required?: boolean } = {}
 ): HookDecorator {
   // tslint:disable-next-line
   const required = options.required ?? true;
+  name = name.toLowerCase();
 
   let validateSchema: ValidateFunction|undefined;
 
@@ -52,13 +51,12 @@ export function ValidateQueryParam(
         type: 'object',
       });
     }
-
-    if (!validateSchema(ctx.request.query)) {
-      return new HttpResponseBadRequest({ query: validateSchema.errors });
+    if (!validateSchema(ctx.request.headers)) {
+      return new HttpResponseBadRequest({ headers: validateSchema.errors });
     }
   }
 
-  const param: IApiQueryParameter = { in: 'query', name };
+  const param: IApiHeaderParameter = { in: 'header', name };
   if (required) {
     param.required = required;
   }
